+++ deadchest-core/src/main/java/me/crylonz/deadchest/utils/Utils.java
@@ -136,17 +136,26 @@ public class Utils {
                 skull.update();
                 break;
             case 3:
-                block.setType(Material.BARREL);
+                block.setType(graveBlockOrChest("BARREL"));
                 break;
             case 4:
-                block.setType(Material.SHULKER_BOX);
+                block.setType(graveBlockOrChest("SHULKER_BOX"));
                 break;
             case 5:
-                block.setType(Material.ENDER_CHEST);
+                block.setType(graveBlockOrChest("ENDER_CHEST"));
                 break;
             default:
                 block.setType(Material.CHEST);
         }
+    }
+
+    /**
+     * @param name material name of the configured grave block
+     * @return the material, or a chest when this Minecraft version does not have it
+     */
+    private static Material graveBlockOrChest(String name) {
+        Material material = RegistryCompat.material(name);
+        return material == null ? Material.CHEST : material;
     }
 
     public static ArmorStand[] createHolograms(Block block, String deathDisplayName) {